Fixing ambiguity in log call in fiona.
git-svn-id: http://svn.seqan.de/seqan/trunk@8817 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/apps/fiona/fiona.cpp
@@ -1568,7 +1568,7 @@ int main(int argc, const char* argv[])
 
 		resize(logCorrections,cycle);
 		resize(roundsDone,cycle);
-		logCorrections[cycle-1] = (double)log(numCorrected);
+		logCorrections[cycle-1] = (double)log((double)numCorrected);
 		roundsDone[cycle-1]     = (double)cycle;
             
 		if(cycle >= 3){
